PhotoRemark/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pydub==0.25.1
websocket-client==1.6.4